pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mnemo-api"
version = "1.0.0"
description = "Mnemo Learning API — spaced repetition and active recall, delivered as a developer API"
authors = [{ name = "Enow Sinke", email = "sinke237@github.com" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.1",
    "pydantic-settings==2.6.1",
    "sqlalchemy==2.0.36",
    "aiosqlite==0.20.0",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "redis==5.2.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.18",
    "pytz==2024.2",
    "pycountry==24.6.1",
    "python-dateutil==2.9.0",
    "httpx==0.27.2",
    "structlog==24.4.0",
    "email-validator==2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "httpx==0.27.2",
    "ruff==0.8.2",
    "black==24.10.0",
    "mypy==1.13.0",
    "types-pytz==2024.2.0.20241003",
    "types-python-dateutil==2.9.0.20241206",
    "factory-boy==3.3.1",
    "faker==33.1.0",
    "freezegun==1.5.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mnemo"]

# ── Ruff (linter) ─────────────────────────────────────────────────────────────
[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit (security)
    "ANN", # flake8-annotations
]
ignore = [
    "S104",   # possible binding to all interfaces (fine for docker)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN", "S106"]  # allow assert, skip annotations, allow test passwords
"src/mnemo/db/redis.py" = ["ANN"]    # redis.asyncio lacks complete type stubs

# ── Black (formatter) ─────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py312"]

# ── Mypy (type checker) ───────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "mnemo.db.redis"
ignore_errors = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["src"]
addopts = "--cov=src/mnemo --cov-report=term-missing --cov-report=xml --cov-fail-under=70"

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/mnemo"]
omit = ["*/migrations/*"]